# expression: implement vectorized evaluation for `builtinTrim3ArgsSig`