Signed-off-by: AerysNan <aerysnan@gmail.com>

PCP #12106 

### What problem does this PR solve? <!--add issue link with summary if exists-->

Implement vectorized evaluation for builtinTrim3ArgsSig.

### What is changed and how it works?

```
BenchmarkVectorizedBuiltinStringFunc/builtinTrim3ArgsSig-VecBuiltinFunc-4                  29697             37926 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinTrim3ArgsSig-NonVecBuiltinFunc-4               17577             76310 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinTrim3ArgsSig-VecBuiltinFunc#01-4               26559             46683 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinTrim3ArgsSig-NonVecBuiltinFunc#01-4            12838             96674 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinTrim3ArgsSig-VecBuiltinFunc#02-4               25411             43928 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinTrim3ArgsSig-NonVecBuiltinFunc#02-4            14025             82855 ns/op               0 B/op          0 allocs/op   
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test